refactor: merge cli + server into a single apps/server app
`packages/server` (the loopback HTTP+WS transport) had exactly one
consumer — the `apps/cli` launcher — so the two were one shippable
"headless server" artifact split across an app and a library for no
reason. Collapse them:

- rename `apps/cli` → `apps/server` (keeps the `inteligir` bin and the
  `@repo/server` package name, freed up by the merge)
- fold the transport in as `apps/server/src/create-server.ts` (+ its test
  suite and vitest config); `main.ts` now imports it relatively
- delete `packages/server`

The server app depends on core + host + app and imports the transport
directly. Lint boundary, knip workspaces, and the run-mode docs are
updated to the new layout.
