+++ README.md
@@ -11,13 +11,12 @@ marketing site.
 ```
 apps/              Shippable artifacts
   desktop/         Thin Electron shell over host + app (@repo/desktop)
-  cli/             `inteligir <vault>`: boot host+server, open the browser (@repo/cli)
+  server/          `inteligir <vault>`: boot host, serve app over loopback HTTP+WS, open the browser (@repo/server)
   web/             TanStack Start marketing site on Cloudflare Workers (landing page only)
 packages/          Libraries
   app/             Portable UI — the whole workspace as a browser React app (@repo/app)
   core/            Isomorphic contract: Bridge/IPC registry, domain schemas (@repo/core)
   host/            Platform-agnostic node backend: vault, pi, delegation, executor, voice (@repo/host)
-  server/          Loopback HTTP+WS host: folds the registry over WS, serves the app build (@repo/server)
   ui/              Shared UI components (@repo/ui)
   agent-runtime/   Filesystem + install primitives for the agent (@repo/agent-runtime)
   pi-driver/       Wrapper around pi-coding-agent — stable surface for app code (@repo/pi-driver)
@@ -28,12 +27,11 @@ Workspace `README.md`s:
 | Workspace                | README                                                                    |
 | ------------------------ | ------------------------------------------------------------------------- |
 | `apps/desktop`           | [Electron shell — process boundary, packaging](./apps/desktop/README.md)  |
-| `apps/cli`               | [`inteligir` launcher](./apps/cli/README.md)                              |
+| `apps/server`            | [`inteligir` launcher + loopback HTTP+WS host](./apps/server/README.md)   |
 | `apps/web`               | [static marketing site](./apps/web/README.md)                             |
 | `packages/app`           | [portable UI — Bridge-injected workspace](./packages/app/README.md)       |
 | `packages/core`          | [isomorphic contract — IPC registry, schemas](./packages/core/README.md)  |
 | `packages/host`          | [node backend — createHost, HostPlatform](./packages/host/README.md)      |
-| `packages/server`        | [browser host — loopback HTTP+WS](./packages/server/README.md)            |
 | `packages/agent-runtime` | [install / seed / run-cli primitives](./packages/agent-runtime/README.md) |
 | `packages/pi-driver`     | [pi-coding-agent wrapper](./packages/pi-driver/README.md)                 |
 | `packages/ui`            | [shared design system](./packages/ui/README.md)                           |
