+++ apps/server/tsconfig.json
@@ -8,10 +8,14 @@
     "paths": {
       "@repo/core/*": ["../../packages/core/src/*"],
       "@repo/host/*": ["../../packages/host/src/*"],
-      "@repo/server/*": ["../../packages/server/src/*"],
       "@repo/pi-driver/*": ["../../packages/pi-driver/src/*"],
       "@repo/agent-runtime/*": ["../../packages/agent-runtime/src/*"]
     }
   },
-  "include": ["src", "bin", "../../packages/host/src/voice/sherpa-onnx-node.d.ts"]
+  "include": [
+    "src",
+    "bin",
+    "vitest.config.ts",
+    "../../packages/host/src/voice/sherpa-onnx-node.d.ts"
+  ]
 }
