+++ knip.json
@@ -42,21 +42,17 @@
       "project": ["src/**/*.{ts,tsx,css,js}", "dev/**/*.{ts,tsx}", "web/**/*.{ts,tsx}"],
       "includeEntryExports": false
     },
-    "apps/cli": {
+    "apps/server": {
       // bin/inteligir.js (package.json bin) and src/main.ts (its import +
-      // the dev script) are auto-detected entries.
+      // the dev script) are auto-detected entries; the folded-in WS transport
+      // (create-server.ts) has its own test suite.
+      "entry": ["src/__tests__/**/*.test.ts"],
       "project": ["src/**/*.ts", "bin/**/*.js"],
-      // @repo/app: never imported — the cli resolves its dist-web assets at
+      // @repo/app: never imported — the server resolves its dist-web assets at
       // runtime via require.resolve("@repo/app/package.json"), and the
       // workspace dep is what makes that resolvable (and, at Phase 7, what
       // ships the build to npm).
       "ignoreDependencies": ["@repo/app"]
-    },
-    "packages/server": {
-      // Consumed by @repo/cli through its `./*` exports.
-      "entry": ["src/__tests__/**/*.test.ts"],
-      "project": ["src/**/*.ts"],
-      "includeEntryExports": false
     },
     "packages/host": {
       // The node backend, source-imported by its shells (desktop today,
