+++ packages/host/README.md
@@ -1,6 +1,6 @@
 # `@repo/host` — the node backend
 
-The platform-agnostic backend library: app lifecycle, the agent singleton, vault, knowledge indexes, delegation, executor, voice, and the Bridge handler map. `createHost(platform, options)` (`src/create-host.ts`) composes it all; a shell (the Electron desktop `apps/desktop`, or `@repo/server` behind the cli) injects a `HostPlatform` (`src/platform.ts`) for everything OS/shell-shaped — native dialogs, secret cipher (keychain or file-key), notifications, packaged-resource paths — folds `host.handlers` into its transport, and forwards `host.events`. Only one real host runs at a time: `lib/host-lock.ts` is a pidfile under `~/.inteligir`. The UI talks to it only via methods declared in the registry (`@repo/core/ipc-registry`). No electron imports anywhere in `src/` — lint-enforced.
+The platform-agnostic backend library: app lifecycle, the agent singleton, vault, knowledge indexes, delegation, executor, voice, and the Bridge handler map. `createHost(platform, options)` (`src/create-host.ts`) composes it all; a shell (the Electron desktop `apps/desktop`, or the browser via `@repo/server`) injects a `HostPlatform` (`src/platform.ts`) for everything OS/shell-shaped — native dialogs, secret cipher (keychain or file-key), notifications, packaged-resource paths — folds `host.handlers` into its transport, and forwards `host.events`. Only one real host runs at a time: `lib/host-lock.ts` is a pidfile under `~/.inteligir`. The UI talks to it only via methods declared in the registry (`@repo/core/ipc-registry`). No electron imports anywhere in `src/` — lint-enforced.
 
 ## State machine — three-part split
